model_app/requirements.txt
transformers==4.44.2
torch
shap==0.45.0
cohere==4.34
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.26.0
pandas==2.2.0
gradio==4.44.0
plotly==5.17.0
huggingface-hub==0.23.2
tokenizers==0.19.1